refactor: [#1298] move UdpTrackerContainer to udp-tracker-core package
