+++ src/container.rs
@@ -13,6 +13,7 @@ use bittorrent_tracker_core::torrent::repository::persisted::DatabasePersistentT
 use bittorrent_tracker_core::whitelist;
 use bittorrent_tracker_core::whitelist::manager::WhitelistManager;
 use bittorrent_tracker_core::whitelist::repository::in_memory::InMemoryWhitelist;
+use bittorrent_udp_tracker_core::container::UdpTrackerContainer;
 use bittorrent_udp_tracker_core::services::banning::BanService;
 use bittorrent_udp_tracker_core::{self, MAX_CONNECTION_ID_ERRORS_PER_IP};
 use tokio::sync::RwLock;
@@ -33,9 +34,11 @@ pub struct AppContainer {
     pub in_memory_torrent_repository: Arc<InMemoryTorrentRepository>,
     pub db_torrent_repository: Arc<DatabasePersistentTorrentRepository>,
     pub torrents_manager: Arc<TorrentsManager>,
+
     // UDP Tracker Core Services
     pub ban_service: Arc<RwLock<BanService>>,
     pub udp_stats_event_sender: Arc<Option<Box<dyn bittorrent_udp_tracker_core::statistics::event::sender::Sender>>>,
+
     // HTTP Tracker Core Services
     pub http_stats_event_sender: Arc<Option<Box<dyn bittorrent_http_tracker_core::statistics::event::sender::Sender>>>,
     pub http_stats_repository: Arc<bittorrent_http_tracker_core::statistics::repository::Repository>,
@@ -84,16 +87,6 @@ impl AppContainer {
             udp_stats_repository: self.udp_stats_repository.clone(),
         }
     }
-}
-
-pub struct UdpTrackerContainer {
-    pub core_config: Arc<Core>,
-    pub udp_tracker_config: Arc<UdpTracker>,
-    pub announce_handler: Arc<AnnounceHandler>,
-    pub scrape_handler: Arc<ScrapeHandler>,
-    pub whitelist_authorization: Arc<whitelist::authorization::WhitelistAuthorization>,
-    pub udp_stats_event_sender: Arc<Option<Box<dyn bittorrent_udp_tracker_core::statistics::event::sender::Sender>>>,
-    pub ban_service: Arc<RwLock<BanService>>,
 }
 
 pub struct HttpApiContainer {
@@ -174,30 +167,5 @@ pub fn initialize_http_api_container(core_config: &Arc<Core>, http_api_config: &
         ban_service: ban_service.clone(),
         http_stats_repository: http_stats_repository.clone(),
         udp_stats_repository: udp_stats_repository.clone(),
-    })
-}
-
-#[must_use]
-pub fn initialize_udt_tracker_container(
-    core_config: &Arc<Core>,
-    udp_tracker_config: &Arc<UdpTracker>,
-) -> Arc<UdpTrackerContainer> {
-    let tracker_core_container = TrackerCoreContainer::initialize(core_config);
-
-    // UDP stats
-    let (udp_stats_event_sender, _udp_stats_repository) =
-        bittorrent_udp_tracker_core::statistics::setup::factory(core_config.tracker_usage_statistics);
-    let udp_stats_event_sender = Arc::new(udp_stats_event_sender);
-
-    let ban_service = Arc::new(RwLock::new(BanService::new(MAX_CONNECTION_ID_ERRORS_PER_IP)));
-
-    Arc::new(UdpTrackerContainer {
-        udp_tracker_config: udp_tracker_config.clone(),
-        core_config: core_config.clone(),
-        announce_handler: tracker_core_container.announce_handler.clone(),
-        scrape_handler: tracker_core_container.scrape_handler.clone(),
-        whitelist_authorization: tracker_core_container.whitelist_authorization.clone(),
-        udp_stats_event_sender: udp_stats_event_sender.clone(),
-        ban_service: ban_service.clone(),
     })
 }
